+++ xmlschema/validators/xsdbase.py
@@ -17,7 +17,8 @@ from ..compat import PY3, unicode_type
 from ..etree import etree_tostring, etree_iselement
 from ..exceptions import XMLSchemaValueError, XMLSchemaTypeError
 from ..qnames import (
-    local_name, get_qname, qname_to_prefixed, XSD_ANNOTATION_TAG, XSD_APPINFO_TAG, XSD_DOCUMENTATION_TAG, XML_LANG
+    local_name, get_qname, qname_to_prefixed, XSD_ANNOTATION_TAG, XSD_APPINFO_TAG,
+    XSD_DOCUMENTATION_TAG, XML_LANG, XSD_ANY_TYPE
 )
 from .exceptions import XMLSchemaParseError, XMLSchemaValidationError
 from .parseutils import (
@@ -213,6 +214,16 @@ class XsdComponent(XsdBaseComponent):
         else:
             return error
 
+    def _parse(self):
+        super(XsdComponent, self)._parse()
+        try:
+            if self.elem[0].tag == XSD_ANNOTATION_TAG:
+                self.annotation = XsdAnnotation(self.elem[0], self.schema)
+            else:
+                self.annotation = None
+        except (TypeError, IndexError):
+            self.annotation = None
+
     def _parse_component(self, elem, required=True, strict=True):
         try:
             return get_xsd_component(elem, required, strict)
@@ -309,7 +320,7 @@ class XsdAnnotation(XsdComponent):
         return True
 
     def _parse(self):
-        super(XsdAnnotation, self)._parse()
+        super(XsdComponent, self)._parse()  # Skip parent class method (that parses also annotations)
         self.appinfo = []
         self.documentation = []
         for child in self.elem:
@@ -325,16 +336,10 @@ class XsdAnnotation(XsdComponent):
                 self.documentation.append(child)
 
 
-class XsdAnnotated(XsdComponent):
-    def _parse(self):
-        super(XsdAnnotated, self)._parse()
-        try:
-            if self.elem[0].tag == XSD_ANNOTATION_TAG:
-                self.annotation = XsdAnnotation(self.elem[0], self.schema)
-            else:
-                self.annotation = None
-        except (TypeError, IndexError):
-            self.annotation = None
+class XsdType(XsdComponent):
+
+    base_type = None
+    derivation = None
 
     @property
     def built(self):
@@ -343,6 +348,58 @@ class XsdAnnotated(XsdComponent):
     @property
     def admitted_tags(self):
         raise NotImplementedError
+
+    @staticmethod
+    def is_simple():
+        raise NotImplementedError
+
+    @staticmethod
+    def is_complex():
+        raise NotImplementedError
+
+    def is_empty(self):
+        raise NotImplementedError
+
+    def is_emptiable(self):
+        raise NotImplementedError
+
+    def has_simple_content(self):
+        raise NotImplementedError
+
+    def has_mixed_content(self):
+        raise NotImplementedError
+
+    def is_element_only(self):
+        raise NotImplementedError
+
+    @property
+    def content_type_label(self):
+        if self.is_empty():
+            return 'empty'
+        elif self.has_simple_content():
+            return 'simple'
+        elif self.is_element_only():
+            return 'element-only'
+        elif self.has_mixed_content():
+            return 'mixed'
+        else:
+            return 'unknown'
+
+    def is_derived(self, other, derivation=None):
+        if other.name == XSD_ANY_TYPE or self.base_type == other:
+            return True if derivation is None else derivation == self.derivation
+        elif self.base_type is not None:
+            return self.base_type.is_derived(other, derivation)
+        else:
+            return False
+
+    def is_subtype(self, qname):
+        if qname == XSD_ANY_TYPE or self.name == qname:
+            return True
+        elif self.base_type is not None:
+            return self.base_type.is_subtype(qname)
+        else:
+            return False
 
 
 class ParticleMixin(object):